refactor(abw): decompose standard article stage builder
