+++ apps/ai-blog-writer/apps/frontend/src/features/staging/components/StandardArticleStageBuilder.tsx
@@ -1,46 +1,25 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
-import { Link } from 'react-router-dom'
+import { useCallback, useState } from 'react'
 import { useAuth, usePermissions } from '../../auth'
-import {
-  generateSocialImageFromFeatured as requestGenerateSocialImageFromFeatured,
-  uploadSocialImage as requestUploadSocialImage,
-} from '../../../shared/images'
 import { SeoEditorPanel } from '../../../shared/seo/components/SeoEditorPanel'
-import { getSchemaPublisherConfig } from '../../../shared/seo/services/schema-publisher-config.service'
-import { buildArticleOgUrl } from '../../../shared/seo/utils/buildArticleOgUrl'
-import {
-  applySeoAiPatch,
-  buildSeoAiSeed,
-  parseSeoAiPatch,
-  type SeoAiTarget,
-} from '../../../shared/seo/services/seo-ai.service'
+import type { SeoSection } from '../../../shared/seo/types'
 import { createEmptySeoSection } from '../../../shared/seo/services/seo-section.service'
-import { resolveEditorAssistModelName } from '../api'
 import type { EditorialStageArticleApi, EditorialStageRoutes } from '../features/editorial-stage-article/types'
 import { useEditorialStageArticleScreenViewModel } from '../features/editorial-stage-article/hooks/useEditorialStageArticleScreenViewModel'
-import { EDITOR_MODEL_OPTIONS, resolveEditorModelName } from '../features/editorial-stage-article/constants'
-import { getLocationDisplayName } from '../features/editorial-stage-article/utils/editorial-stage-view.utils'
-import {
-  buildStandardArticleContext,
-  buildLegacyStandardArticleStructuredDataTemplate,
-  buildStandardArticleSeoAiPrompt,
-  shouldAutoManageStandardArticleStructuredData,
-  buildStandardArticleStructuredDataTemplate,
-  isSeoCoreComplete,
-  serializeStandardArticleStructuredDataTemplate,
-  validateStandardArticleSeoSection,
-} from '../features/editorial-stage-article/services/standard-article-seo.service'
-import {
-  buildPrimaryLocationUpdate,
-  getSharedNeighborhoodOptions,
-  sanitizeSharedNeighborhoods,
-} from '../features/editorial-stage-article/utils/sharedNeighborhoods'
+import { useStandardArticleDerivedState } from '../features/editorial-stage-article/hooks/useStandardArticleDerivedState'
+import { useStandardArticleExpansion } from '../features/editorial-stage-article/hooks/useStandardArticleExpansion'
+import { useStandardArticlePermalink } from '../features/editorial-stage-article/hooks/useStandardArticlePermalink'
+import { useStandardArticleSeoActions } from '../features/editorial-stage-article/hooks/useStandardArticleSeoActions'
+import { useStandardArticleStageActions } from '../features/editorial-stage-article/hooks/useStandardArticleStageActions'
+import { useStandardArticleStructuredData } from '../features/editorial-stage-article/hooks/useStandardArticleStructuredData'
 import { FeaturedImageModal } from './editorial-stage/FeaturedImageModal'
 import { BlockImageModal } from './editorial-stage/BlockImageModal'
-import { EditorialTimelineList } from './editorial-stage/EditorialTimelineList'
 import { ArticleExpansionModal } from '../../youtube2blog/components/ArticleExpansionModal'
-import { useArticleExpansion } from '../../youtube2blog/hooks/useArticleExpansion'
-import { parseMarkdownToBlocks } from '../features/editorial-stage-article/workflow.service'
+import { StandardArticleBuilderSidebar } from './standard-article-stage/StandardArticleBuilderSidebar'
+import { StandardArticleContentPanel } from './standard-article-stage/StandardArticleContentPanel'
+import { StandardArticleFeaturedImagePanel } from './standard-article-stage/StandardArticleFeaturedImagePanel'
+import { StandardArticleSetupPanel } from './standard-article-stage/StandardArticleSetupPanel'
+import { StandardArticleStageHero } from './standard-article-stage/StandardArticleStageHero'
+import { StandardArticleStatusBanners } from './standard-article-stage/StandardArticleStatusBanners'
 import '../../singleTypeListicles/styles.css'
 import './standard-article-stage-builder.css'
 
@@ -54,33 +33,6 @@ type StandardArticleStageBuilderProps = {
   syncBehavior?: 'finalize' | 'draft-sync'
   backToStageLabel?: string
 }
-
-function validateSetupStep(title: string, locationId?: number): string[] {
-  const issues: string[] = []
-  if (!title.trim()) issues.push('Step 1 requires an article title.')
-  if (!locationId) issues.push('Step 1 requires a location.')
-  return issues
-}
-
-function validateFeaturedImageStep(featuredImageId?: number): string[] {
-  return featuredImageId ? [] : ['Select a featured image before syncing to Payload.']
-}
-
-function validateContentStep(input: {
-  bodyMarkdown: string
-  editorialBlockingMessages: string[]
-}): string[] {
-  const issues: string[] = []
-  if (!input.bodyMarkdown.trim()) {
-    issues.push('Step 3 requires at least one text block with content.')
-  }
-  if (input.editorialBlockingMessages.length > 0) {
-    issues.push(input.editorialBlockingMessages[0])
-  }
-  return issues
-}
-
-const schemaPublisherConfig = getSchemaPublisherConfig()
 
 export function StandardArticleStageBuilder({
   storageKey,
@@ -96,13 +48,6 @@ export function StandardArticleStageBuilder({
   const { canManagePublished, role } = usePermissions()
   const [localError, setLocalError] = useState<string | null>(null)
   const [localResult, setLocalResult] = useState<string | null>(null)
-  const [isGeneratingSeoTarget, setIsGeneratingSeoTarget] = useState<SeoAiTarget | null>(null)
-  const [isGeneratingSeoImage, setIsGeneratingSeoImage] = useState(false)
-  const [isUploadingOgImage, setIsUploadingOgImage] = useState(false)
-  const [isGeneratingSlug, setIsGeneratingSlug] = useState(false)
-  const lastAutoStructuredDataRef = useRef<string>('')
-
-  const [isExpansionModalOpen, setIsExpansionModalOpen] = useState(false)
 
   const viewModel = useEditorialStageArticleScreenViewModel({
     storageKey,
@@ -111,132 +56,20 @@ export function StandardArticleStageBuilder({
     api,
     token,
   })
-
-  const { status, layout, timelineListProps, sidebarProps, featuredModalProps, blockModalProps } = viewModel
+  const {
+    status,
+    layout,
+    timelineListProps,
+    sidebarProps,
+    featuredModalProps,
+    blockModalProps,
+  } = viewModel
   const stagedArticle = layout?.stagedArticle
-  const seoSection = stagedArticle?.seoSection ?? createEmptySeoSection()
-  const selectedLocation = useMemo(
-    () => sidebarProps?.locations.find((location) => location.id === stagedArticle?.locationId),
-    [sidebarProps?.locations, stagedArticle?.locationId],
-  )
-  const sharedNeighborhoodOptions = useMemo(
-    () => getSharedNeighborhoodOptions(sidebarProps?.locations ?? [], stagedArticle?.locationId),
-    [sidebarProps?.locations, stagedArticle?.locationId],
-  )
-  const selectedSharedNeighborhoods = useMemo(
-    () => sanitizeSharedNeighborhoods(
-      stagedArticle?.sharedNeighborhoods ?? [],
-      sidebarProps?.locations ?? [],
-      stagedArticle?.locationId
-    ),
-    [sidebarProps?.locations, stagedArticle?.locationId, stagedArticle?.sharedNeighborhoods],
-  )
-  const selectedLocationLabel = useMemo(
-    () => getLocationDisplayName(selectedLocation),
-    [selectedLocation],
-  )
-  const bodyMarkdown = useMemo(
-    () => (stagedArticle ? buildStandardArticleContext(stagedArticle) : ''),
-    [stagedArticle],
-  )
+  const derived = useStandardArticleDerivedState(stagedArticle, sidebarProps)
 
-  const step1Issues = useMemo(
-    () => validateSetupStep(stagedArticle?.title || '', stagedArticle?.locationId),
-    [stagedArticle?.title, stagedArticle?.locationId],
-  )
-  const step2Issues = useMemo(
-    () => validateFeaturedImageStep(stagedArticle?.featuredImageId),
-    [stagedArticle?.featuredImageId],
-  )
-  const step3Issues = useMemo(
-    () => validateContentStep({
-      bodyMarkdown,
-      editorialBlockingMessages: sidebarProps?.editorialBlockingMessages || [],
-    }),
-    [bodyMarkdown, sidebarProps?.editorialBlockingMessages],
-  )
-  const seoIssues = useMemo(
-    () => validateStandardArticleSeoSection({
-      seoSection,
-      locationLabel: selectedLocationLabel || undefined,
-    }),
-    [seoSection, selectedLocationLabel],
-  )
-  const featuredImageId = stagedArticle?.featuredImageId
-  const featuredImagePreviewUrl = sidebarProps?.selectedFeaturedImage
-    ? sidebarProps.getImageUrl(sidebarProps.selectedFeaturedImage)
-    : undefined
-  const featuredImageTriggerLabel = featuredImageId
-    ? sidebarProps?.selectedFeaturedImage?.filename || `Image #${featuredImageId} selected`
-    : 'Select Featured Image...'
-  const headerPreviewTitle = stagedArticle?.title.trim() || 'Your article headline will appear here'
-  const isFeaturedImagePlaceholder = !featuredImageId
-  const isPublished = stagedArticle?.payloadStatus === 'published'
-  const isCityPrimaryLocation = selectedLocation?.level === 'city'
-
-  const isStep1Locked = Boolean(stagedArticle?.step1_complete && !stagedArticle?.in_update_mode)
-  const isStep2Locked = Boolean(stagedArticle?.step2_complete && !stagedArticle?.step2_in_update_mode)
-  const isStep3Locked = Boolean(stagedArticle?.step3_complete && !stagedArticle?.step3_in_update_mode)
-  const isSynced = Boolean(stagedArticle?.publishedToPayload)
-  const hasUnsyncedPayloadChanges = Boolean(isSynced && stagedArticle?.hasUnsyncedPayloadChanges)
-  const seoCoreComplete = isSeoCoreComplete(seoSection)
-  const completionPercent = Math.round(([
-    isStep1Locked,
-    isStep3Locked,
-    seoCoreComplete,
-  ].filter(Boolean).length / 3) * 100)
-
-  const syncIssues = useMemo(() => {
-    const issues: string[] = []
-    if (isSynced) {
-      if (step1Issues.length > 0) issues.push(step1Issues[0])
-      if (!stagedArticle?.payloadSlug?.trim()) issues.push('Slug is required before syncing to Payload.')
-      if (step2Issues.length > 0) issues.push(step2Issues[0])
-      if (step3Issues.length > 0) issues.push(step3Issues[0])
-      if (!seoCoreComplete) issues.push('Add SEO title and meta description before syncing to Payload.')
-      if (seoIssues.length > 0) issues.push(seoIssues[0])
-    } else {
-      if (!isStep1Locked) issues.push('Lock Step 1 before syncing to Payload.')
-      if (!stagedArticle?.payloadSlug?.trim()) issues.push('Slug is required before syncing to Payload.')
-      if (step2Issues.length > 0) issues.push(step2Issues[0])
-      if (!isStep3Locked) issues.push('Lock Step 3 before syncing to Payload.')
-      if (!seoCoreComplete) issues.push('Add SEO title and meta description before syncing to Payload.')
-      if (step3Issues.length > 0) issues.push(step3Issues[0])
-      if (seoIssues.length > 0) issues.push(seoIssues[0])
-    }
-    return issues
-  }, [isSynced, isStep1Locked, isStep3Locked, seoCoreComplete, seoIssues, stagedArticle?.payloadSlug, step1Issues, step2Issues, step3Issues])
-
-  const setStageArticle = useCallback((updates: Partial<NonNullable<typeof stagedArticle>>) => {
-    sidebarProps?.onUpdateStagedArticle(updates)
-  }, [sidebarProps])
-
-  const expansion = useArticleExpansion(stagedArticle?.runId ?? null)
-
-  const handleOpenDeepExpand = useCallback(() => {
-    if (!stagedArticle) return
-    expansion.reset()
-    expansion.startExpansion(
-      stagedArticle.content,
-      stagedArticle.originalType,
-      stagedArticle.title || stagedArticle.originalTitle,
-    )
-    setIsExpansionModalOpen(true)
-  }, [expansion, stagedArticle])
-
-  const handleAcceptExpansion = useCallback((expandedMarkdown: string) => {
-    const blocks = parseMarkdownToBlocks(expandedMarkdown)
-    setStageArticle({ blocks, content: expandedMarkdown, lexicalConverted: false })
-    expansion.reset()
-    setIsExpansionModalOpen(false)
-  }, [expansion, setStageArticle])
-
-  const handleCloseExpansion = useCallback(() => {
-    expansion.reset()
-    setIsExpansionModalOpen(false)
-  }, [expansion])
-
-  const updateSeoSection = useCallback((next: typeof seoSection | ((current: typeof seoSection) => typeof seoSection)) => {
+  const updateSeoSection = useCallback((
+    next: SeoSection | ((current: SeoSection) => SeoSection),
+  ) => {
     if (!stagedArticle || !sidebarProps) return
     const resolved = typeof next === 'function'
       ? next(stagedArticle.seoSection ?? createEmptySeoSection())
@@ -244,271 +77,57 @@ export function StandardArticleStageBuilder({
     sidebarProps.onUpdateStagedArticle({ seoSection: resolved })
   }, [sidebarProps, stagedArticle])
 
-  useEffect(() => {
-    if (!stagedArticle || !isStep3Locked) return
-
-    const nextStructuredData = serializeStandardArticleStructuredDataTemplate(
-      buildStandardArticleStructuredDataTemplate({
-        stagedArticle: {
-          ...stagedArticle,
-          seoSection,
-        },
-        locationLabel: selectedLocationLabel || undefined,
-        publisherConfig: schemaPublisherConfig,
-      }),
-    )
-    const legacyStructuredData = serializeStandardArticleStructuredDataTemplate(
-      buildLegacyStandardArticleStructuredDataTemplate({
-        stagedArticle: {
-          ...stagedArticle,
-          seoSection,
-        },
-        locationLabel: selectedLocationLabel || undefined,
-      }),
-    )
-
-    const existingStructuredData = seoSection.structuredData.trim()
-    const lastAutoStructuredData = lastAutoStructuredDataRef.current.trim()
-    const isAutoManaged = shouldAutoManageStandardArticleStructuredData({
-      existingStructuredData,
-      lastAutoStructuredData,
-      nextStructuredData,
-      legacyStructuredData,
-    })
-
-    if (!isAutoManaged) return
-    if (existingStructuredData === nextStructuredData) {
-      lastAutoStructuredDataRef.current = nextStructuredData
-      return
-    }
-
-    lastAutoStructuredDataRef.current = nextStructuredData
-    updateSeoSection((current) => ({
-      ...current,
-      structuredData: nextStructuredData,
-    }))
-  }, [
-    isStep3Locked,
-    seoSection,
-    selectedLocationLabel,
+  const actions = useStandardArticleStageActions({
     stagedArticle,
+    sidebarProps,
+    step1Issues: derived.step1Issues,
+    step3Issues: derived.step3Issues,
+    syncIssues: derived.syncIssues,
+    setLocalError,
+    setLocalResult,
+  })
+  const permalink = useStandardArticlePermalink({
+    api,
+    stagedArticle,
+    selectedLocation: derived.selectedLocation,
     updateSeoSection,
-  ])
+    setLocalError,
+  })
+  const seo = useStandardArticleSeoActions({
+    api,
+    token,
+    stagedArticle,
+    featureLabel,
+    selectedLocationLabel: derived.selectedLocationLabel,
+    seoSection: derived.seoSection,
+    updateSeoSection,
+    setLocalError,
+    setLocalResult,
+  })
+  const expansion = useStandardArticleExpansion(stagedArticle, actions.setStageArticle)
 
-  const handleAutoFillOgUrl = useCallback(() => {
-    const slug = stagedArticle?.payloadSlug?.trim()
-    const country = selectedLocation?.country
-    const city = selectedLocation?.city
-    if (!slug || !country) return
-    const url = buildArticleOgUrl(country, city, null, slug)
-    if (!url) return
-    updateSeoSection((current) => ({ ...current, openGraph: { ...current.openGraph, url } }))
-  }, [stagedArticle?.payloadSlug, selectedLocation, updateSeoSection])
+  useStandardArticleStructuredData({
+    stagedArticle,
+    seoSection: derived.seoSection,
+    selectedLocationLabel: derived.selectedLocationLabel,
+    enabled: derived.isStep3Locked,
+    updateSeoSection,
+  })
 
-  const handleContinueSetup = useCallback(() => {
-    if (!stagedArticle || !sidebarProps) return
-    if (step1Issues.length > 0) {
-      setLocalError(step1Issues[0])
-      return
-    }
-    if (!stagedArticle.payloadSlug?.trim()) {
-      setLocalError('Slug is required before continuing.')
-      return
-    }
-    setLocalError(null)
-    setLocalResult(null)
-    sidebarProps.onUpdateStagedArticle({
-      step1_complete: true,
-      in_update_mode: false,
-      step2_complete: false,
-      step2_in_update_mode: false,
-      step3_complete: false,
-      step3_in_update_mode: false,
-    })
-  }, [sidebarProps, stagedArticle, step1Issues])
+  const generateSlug = useCallback(async () => {
+    const slug = await permalink.generateSlug()
+    if (slug) actions.setStageArticle({ payloadSlug: slug })
+  }, [actions, permalink])
 
-  const handleGenerateSlugWithAi = useCallback(async () => {
-    if (!stagedArticle?.title.trim() || !sidebarProps) return
-    setIsGeneratingSlug(true)
-    try {
-      const response = await api.rewriteBlockWithAi({
-        prompt: `Generate a clean SEO-friendly URL slug for this article title:\n\nTitle: ${stagedArticle.title.trim()}\n\nRules:\n- Think like a real user searching Google.\n- Keep the most important search keywords.\n- Remove filler words like "the," "a," "an," "in," "of," "to," and "for" unless they are needed.\n- Keep it short, readable, and specific.\n- Use lowercase only.\n- Use hyphens between words.\n- Do not keyword-stuff.\n- Do not add words that are not strongly related to the title.\n- Prefer search-intent wording over matching the title exactly.\n- Return only the slug, no explanation.\n\nExample:\nTitle: The Best Steakhouses in Las Vegas\nSlug: best-steakhouses-las-vegas`,
-        blockContent: stagedArticle.title.trim(),
-        articleTitle: stagedArticle.title.trim(),
-      })
-      const slug = response.rewritten_content?.trim()
-      if (slug) {
-        sidebarProps.onUpdateStagedArticle({ payloadSlug: slug })
-      }
-    } catch (err) {
-      setLocalError(err instanceof Error ? err.message : 'Failed to generate slug with AI.')
-    } finally {
-      setIsGeneratingSlug(false)
-    }
-  }, [api, sidebarProps, stagedArticle])
-
-  const handleContinueFeaturedImage = useCallback(() => {
-    if (!sidebarProps) return
-    setLocalError(null)
-    setLocalResult(null)
-    sidebarProps.onUpdateStagedArticle({
-      step2_complete: true,
-      step2_in_update_mode: false,
-    })
-  }, [sidebarProps])
-
-  const handleContinueContent = useCallback(() => {
-    if (!sidebarProps) return
-    if (step3Issues.length > 0) {
-      setLocalError(step3Issues[0])
-      return
-    }
-    setLocalError(null)
-    setLocalResult(null)
-    sidebarProps.onUpdateStagedArticle({
-      step3_complete: true,
-      step3_in_update_mode: false,
-    })
-  }, [sidebarProps, step3Issues])
-
-  const handleGenerateSeoWithAi = useCallback(async (target: SeoAiTarget = 'all') => {
-    if (!stagedArticle) return
-
-    const articleContext = buildStandardArticleContext(stagedArticle).trim()
-    const articleTitle = stagedArticle.title.trim()
-    const hasSourceContent = Boolean(articleTitle || articleContext)
-    if (!hasSourceContent) {
-      setLocalError('Add article content before generating SEO with AI.')
-      return
-    }
-
-    setLocalError(null)
-    setLocalResult(null)
-    setIsGeneratingSeoTarget(target)
-
-    try {
-      const response = await api.generateSeoMetadataWithAi({
-        prompt: buildStandardArticleSeoAiPrompt({
-          location: selectedLocationLabel || undefined,
-          target,
-        }),
-        seed: buildSeoAiSeed(seoSection),
-        modelName: resolveEditorAssistModelName(stagedArticle.editorModelName),
-        articleTitle,
-        articleContext: articleContext || undefined,
-      })
-
-      if (!response.seo_patch || Object.keys(response.seo_patch).length === 0) {
-        throw new Error('AI returned an empty SEO patch.')
-      }
-
-      // The patch arrives schema-validated from the forced tool call; parse
-      // still applies length clipping and URL validation.
-      const seoPatch = parseSeoAiPatch(JSON.stringify(response.seo_patch))
-      updateSeoSection((current) => applySeoAiPatch(current, seoPatch, target))
-      setLocalResult(`Updated ${target === 'all' ? 'SEO metadata' : target} with AI.`)
-    } catch (err) {
-      setLocalError(err instanceof Error ? err.message : 'Failed to generate SEO with AI.')
-    } finally {
-      setIsGeneratingSeoTarget(null)
-    }
-  }, [api, seoSection, selectedLocationLabel, stagedArticle, updateSeoSection])
-
-  const handleGenerateSeoImageFromFeatured = useCallback(async () => {
-    if (!token || !stagedArticle?.featuredImageId) {
-      setLocalError('Select a featured image before generating an OG image.')
-      return
-    }
-
-    setLocalError(null)
-    setLocalResult(null)
-    setIsGeneratingSeoImage(true)
-
-    try {
-      const response = await requestGenerateSocialImageFromFeatured(
-        { featuredAssetId: stagedArticle.featuredImageId },
-        token,
-      )
-      const socialImageUrl = response.generatedImageUrl.trim()
-      if (!socialImageUrl) {
-        throw new Error('Social image generation did not return a usable URL.')
-      }
-
-      updateSeoSection((current) => ({
-        ...current,
-        openGraph: {
-          ...current.openGraph,
-          imageUrl: socialImageUrl,
-        },
-        twitterCard: {
-          ...current.twitterCard,
-          imageUrl: socialImageUrl,
-        },
-      }))
-      setLocalResult('Generated a 1200x630 social image from the featured image.')
-    } catch (err) {
-      setLocalError(err instanceof Error ? err.message : 'Failed to generate social image.')
-    } finally {
-      setIsGeneratingSeoImage(false)
-    }
-  }, [stagedArticle?.featuredImageId, token, updateSeoSection])
-
-  const handleUploadOgImageFile = useCallback(async (file: File) => {
-    if (!token || !stagedArticle?.locationId) {
-      throw new Error('Select a location before uploading an OG image.')
-    }
-
-    setLocalError(null)
-    setLocalResult(null)
-    setIsUploadingOgImage(true)
-
-    try {
-      const uploadResponse = await requestUploadSocialImage(
-        file,
-        stagedArticle.title.trim() || `${featureLabel} social image`,
-        stagedArticle.locationId,
-        token,
-        '',
-      )
-
-      const socialImageUrl = uploadResponse.generatedImageUrl.trim()
-      if (!socialImageUrl) {
-        throw new Error('OG image upload did not return a usable URL.')
-      }
-
-      updateSeoSection((current) => ({
-        ...current,
-        openGraph: {
-          ...current.openGraph,
-          imageUrl: socialImageUrl,
-        },
-        twitterCard: {
-          ...current.twitterCard,
-          imageUrl: socialImageUrl,
-        },
-      }))
-      setLocalResult('Uploaded and applied a custom OG image.')
-    } catch (err) {
-      const message = err instanceof Error ? err.message : 'Failed to upload OG image.'
-      setLocalError(message)
-      throw err instanceof Error ? err : new Error(message)
-    } finally {
-      setIsUploadingOgImage(false)
-    }
-  }, [featureLabel, stagedArticle?.locationId, stagedArticle?.title, token, updateSeoSection])
-
-  const handlePayloadSubmit = useCallback((targetStatus: 'draft' | 'published') => {
-    if (!sidebarProps) return
-    if (syncIssues.length > 0) {
-      setLocalError(syncIssues[0])
-      return
-    }
-    setLocalError(null)
-    setLocalResult(null)
-    sidebarProps.onPublish(targetStatus)
-  }, [sidebarProps, syncIssues])
-
-  if (status.isLoading || !stagedArticle || !layout || !timelineListProps || !sidebarProps || !featuredModalProps || !blockModalProps) {
+  if (
+    status.isLoading
+    || !stagedArticle
+    || !layout
+    || !timelineListProps
+    || !sidebarProps
+    || !featuredModalProps
+    || !blockModalProps
+  ) {
     return (
       <div className="stl-page stl-single-type-page sab-stage-page">
         <p className="stl-placeholder">{status.error ? status.error : 'Loading builder...'}</p>
@@ -516,472 +135,130 @@ export function StandardArticleStageBuilder({
     )
   }
 
+  const featuredImagePreviewUrl = sidebarProps.selectedFeaturedImage
+    ? sidebarProps.getImageUrl(sidebarProps.selectedFeaturedImage)
+    : undefined
+  const featuredImageTriggerLabel = stagedArticle.featuredImageId
+    ? sidebarProps.selectedFeaturedImage?.filename || `Image #${stagedArticle.featuredImageId} selected`
+    : 'Select Featured Image...'
+
   return (
     <>
       <div className="stl-page stl-single-type-page sab-stage-page">
-        <header className="stl-hero sab-stage-hero">
-          <div>
-            <p className="stl-eyebrow">{heroEyebrow}</p>
-            <h1>{stagedArticle.payloadArticleId ? `Sync Draft #${stagedArticle.payloadArticleId}` : 'Stage Article'}</h1>
-            <p className="stl-lede">{heroDescription}</p>
-            <div className="sab-stage-hero-meta">
-              {stagedArticle.originalType ? <span className="sab-stage-pill">{stagedArticle.originalType}</span> : null}
-              {stagedArticle.payloadArticleId ? <span className="sab-stage-pill">Linked Payload #{stagedArticle.payloadArticleId}</span> : null}
-              {!isSynced ? <span className="sab-stage-pill">{completionPercent}% ready</span> : null}
-            </div>
-          </div>
-          <div className="stl-hero-actions">
-            <Link to={layout.stagePath} className="stl-btn stl-btn-secondary">{backToStageLabel}</Link>
-            <button type="button" className="stl-btn stl-btn-danger" onClick={layout.onDelete}>
-              Delete Staged Article
-            </button>
-          </div>
-        </header>
+        <StandardArticleStageHero
+          stagedArticle={stagedArticle}
+          eyebrow={heroEyebrow}
+          description={heroDescription}
+          backLabel={backToStageLabel}
+          stagePath={layout.stagePath}
+          completionPercent={derived.completionPercent}
+          isSynced={derived.isSynced}
+          onDelete={layout.onDelete}
+        />
 
         <div className="stl-builder-layout">
           <main className="stl-builder-main">
-            {status.saveConflict ? (
-              <div className="stl-summary-warning" role="alert">
-                <strong>Draft changed elsewhere.</strong>
-                <p>
-                  This draft was modified
-                  {status.saveConflict.current?.lastEditedBy?.email
-                    ? ` by ${status.saveConflict.current.lastEditedBy.email}`
-                    : status.saveConflict.current === null
-                      ? ' (it was deleted)'
-                      : ' by another user'}
-                  {' '}since you opened it. Autosave is paused — reload to continue; unsynced local changes will be lost.
-                </p>
-                <button type="button" className="stl-btn stl-btn-secondary" onClick={() => window.location.reload()}>
-                  Reload draft
-                </button>
-              </div>
-            ) : null}
-            {localError ? <p className="stl-error">{localError}</p> : null}
-            {localResult ? <p className="stl-success">{localResult}</p> : null}
+            <StandardArticleStatusBanners
+              status={status}
+              localError={localError}
+              localResult={localResult}
+              hasUnsyncedPayloadChanges={derived.hasUnsyncedPayloadChanges}
+              isPublished={derived.isPublished}
+              isPublishing={sidebarProps.isPublishing}
+              syncIssues={derived.syncIssues}
+              onSubmit={actions.submitToPayload}
+            />
 
-            {hasUnsyncedPayloadChanges ? (
-              <div className="stl-out-of-sync-banner" role="status">
-                <span className="stl-out-of-sync-banner__dot" aria-hidden="true" />
-                <span className="stl-out-of-sync-banner__text">
-                  Out of sync — you have local changes. Sync to Payload to apply them to the live article.
-                </span>
-                <button
-                  type="button"
-                  className="stl-btn stl-out-of-sync-banner__btn"
-                  onClick={() => handlePayloadSubmit(isPublished ? 'published' : 'draft')}
-                  disabled={sidebarProps.isPublishing || syncIssues.length > 0}
-                >
-                  {sidebarProps.isPublishing ? 'Syncing...' : 'Save & Sync'}
-                </button>
-              </div>
-            ) : null}
+            <StandardArticleSetupPanel
+              stagedArticle={stagedArticle}
+              locations={sidebarProps.locations}
+              sharedNeighborhoodOptions={derived.sharedNeighborhoodOptions}
+              selectedSharedNeighborhoods={derived.selectedSharedNeighborhoods}
+              isCityPrimaryLocation={derived.selectedLocation?.level === 'city'}
+              isSynced={derived.isSynced}
+              isStep1Locked={derived.isStep1Locked}
+              isGeneratingSlug={permalink.isGeneratingSlug}
+              onUpdateTitle={layout.onUpdateTitle}
+              onUpdateArticle={actions.setStageArticle}
+              onContinue={actions.continueSetup}
+              onGenerateSlug={generateSlug}
+            />
 
-            <section className="stl-panel sab-stage-panel">
-              <div className="stl-panel-header">
-                <h2>{!isSynced ? <span className="stl-kicker">Step 1</span> : null} Setup</h2>
-                {!isSynced ? (
-                  <div className="stl-inline-actions">
-                    {isStep1Locked ? (
-                      stagedArticle.in_update_mode ? (
-                        <>
-                          <button type="button" className="stl-btn stl-btn-secondary" onClick={() => setStageArticle({ in_update_mode: false })}>
-                            Cancel
-                          </button>
-                          <button type="button" className="stl-btn" onClick={handleContinueSetup}>
-                            Save Setup
-                          </button>
-                        </>
-                      ) : (
-                        <button type="button" className="stl-btn stl-btn-secondary" onClick={() => setStageArticle({ in_update_mode: true })}>
-                          Update Setup
-                        </button>
-                      )
-                    ) : (
-                      <button type="button" className="stl-btn" onClick={handleContinueSetup}>
-                        Continue to Step 2
-                      </button>
-                    )}
-                  </div>
-                ) : null}
-              </div>
-
-              <div className="sab-stage-field-grid">
-                <label className="stl-field">
-                  <span>Article Title</span>
-                  <input
-                    value={stagedArticle.title}
-                    onChange={(event) => layout.onUpdateTitle(event.target.value)}
-                    placeholder="Title the article draft"
-                  />
-                </label>
-
-                <label className="stl-field">
-                  <span>Location</span>
-                  <select
-                    value={stagedArticle.locationId || ''}
-                    onChange={(event) => {
-                      const nextLocationId = Number(event.target.value) || undefined
-                      sidebarProps.onUpdateStagedArticle(buildPrimaryLocationUpdate({
-                        locations: sidebarProps.locations,
-                        nextLocationId,
-                        sharedNeighborhoods: stagedArticle.sharedNeighborhoods,
-                      }))
-                    }}
-                  >
-                    <option value="">Select location</option>
-                    {sidebarProps.locations.map((location) => (
-                      <option key={location.id} value={location.id}>
-                        {getLocationDisplayName(location)}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-
-                {isCityPrimaryLocation ? (
-                  <label className="stl-field">
-                    <span>Shared Neighborhoods</span>
-                    <small>Optional. Exact neighborhood scoping only.</small>
-                    <select
-                      multiple
-                      value={selectedSharedNeighborhoods.map(String)}
-                      onChange={(event) => {
-                        const nextSharedNeighborhoods = Array.from(
-                          event.currentTarget.selectedOptions,
-                          (option) => Number(option.value)
-                        ).filter((value) => Number.isFinite(value) && value > 0)
-
-                        sidebarProps.onUpdateStagedArticle({
-                          sharedNeighborhoods: sanitizeSharedNeighborhoods(
-                            nextSharedNeighborhoods,
-                            sidebarProps.locations,
-                            stagedArticle.locationId
-                          ),
-                        })
-                      }}
-                      style={{ minHeight: '8rem' }}
-                      disabled={sharedNeighborhoodOptions.length === 0}
-                    >
-                      {sharedNeighborhoodOptions.map((location) => (
-                        <option key={location.id} value={location.id}>
-                          {getLocationDisplayName(location)}
-                        </option>
-                      ))}
-                    </select>
-                    {sharedNeighborhoodOptions.length === 0 ? (
-                      <small>No neighborhoods are available for this city yet.</small>
-                    ) : null}
-                  </label>
-                ) : null}
-
-                <label className="stl-field">
-                  <span>AI Model</span>
-                  <select
-                    value={resolveEditorModelName(stagedArticle.editorModelName)}
-                    onChange={(event) => sidebarProps.onUpdateStagedArticle({
-                      editorModelName: resolveEditorModelName(event.target.value),
-                    })}
-                  >
-                    {EDITOR_MODEL_OPTIONS.map((modelOption) => (
-                      <option key={modelOption.value} value={modelOption.value}>
-                        {modelOption.label}
-                      </option>
-                    ))}
-                  </select>
-                </label>
-
-                <label className="stl-field">
-                  <span>Slug *</span>
-                  <small>URL-friendly identifier (e.g. medellin-digital-nomad-guide-2026)</small>
-                  <div className="stl-seo-input-wrap">
-                    <input
-                      className="stl-seo-input-with-ai"
-                      value={stagedArticle.payloadSlug || ''}
-                      onChange={(event) => {
-                        const slug = event.target.value
-                        sidebarProps.onUpdateStagedArticle({ payloadSlug: slug })
-                      }}
-                      placeholder="e.g. best-steakhouses-las-vegas"
-                      disabled={!isSynced && isStep1Locked}
-                    />
-                    {(isSynced || !isStep1Locked) ? (
-                      <span className="stl-seo-ai-trigger-wrap">
-                        <button
-                          type="button"
-                          className="stl-btn stl-btn-secondary stl-seo-ai-btn"
-                          onClick={() => void handleGenerateSlugWithAi()}
-                          disabled={isGeneratingSlug || !stagedArticle.title.trim()}
-                        >
-                          {isGeneratingSlug ? 'Generating...' : 'AI'}
-                        </button>
-                      </span>
-                    ) : null}
-                  </div>
-                </label>
-              </div>
-
-              {isStep1Locked && !isSynced ? (
-                <p className="sab-stage-summary">
-                  Locked with title, primary location, optional shared neighborhoods, and AI model. Updating setup will unlock later steps.
-                </p>
-              ) : null}
-            </section>
-
-            {(isStep1Locked || isSynced) ? (
-              <section className="stl-panel sab-stage-panel">
-                <div className="stl-panel-header">
-                  <h2>{!isSynced ? <span className="stl-kicker">Step 2</span> : null} Featured Image</h2>
-                  {!isSynced ? (
-                    <div className="stl-inline-actions">
-                      {isStep2Locked ? (
-                        stagedArticle.step2_in_update_mode ? (
-                          <>
-                            <button type="button" className="stl-btn stl-btn-secondary" onClick={() => setStageArticle({ step2_in_update_mode: false })}>
-                              Cancel
-                            </button>
-                            <button type="button" className="stl-btn" onClick={handleContinueFeaturedImage}>
-                              Save Featured Image
-                            </button>
-                          </>
-                        ) : (
-                          <button type="button" className="stl-btn stl-btn-secondary" onClick={() => setStageArticle({ step2_in_update_mode: true })}>
-                            Update Featured Image
-                          </button>
-                        )
-                      ) : (
-                        <button type="button" className="stl-btn" onClick={handleContinueFeaturedImage}>
-                          Continue to Step 3
-                        </button>
-                      )}
-                    </div>
-                  ) : null}
-                </div>
-
-                <fieldset className="stl-panel-fieldset" disabled={!isSynced && isStep2Locked}>
-                  <div className="stl-field">
-                    <span>Featured Image</span>
-                    {!featuredImageId ? (
-                      <button
-                        type="button"
-                        className="stl-picker-trigger"
-                        onClick={sidebarProps.onOpenFeaturedImageModal}
-                      >
-                        <span className="stl-picker-trigger__preview">
-                          <span className={`stl-picker-trigger__label${isFeaturedImagePlaceholder ? ' stl-picker-trigger__label--placeholder' : ''}`}>
-                            {featuredImageTriggerLabel}
-                          </span>
-                        </span>
-                        <span className="stl-picker-trigger__caret">▼</span>
-                      </button>
-                    ) : (
-                      <div className="stl-featured-header-preview">
-                        <button
-                          type="button"
-                          className="stl-featured-header-preview__media"
-                          onClick={sidebarProps.onOpenFeaturedImageModal}
-                        >
-                          {featuredImagePreviewUrl ? (
-                            <img src={featuredImagePreviewUrl} alt="" />
-                          ) : (
-                            <div className="stl-featured-header-preview__fallback">Image selected</div>
-                          )}
-                          <div className="stl-featured-header-preview__overlay">
-                            <p className="stl-featured-header-preview__title">{headerPreviewTitle}</p>
-                          </div>
-                        </button>
-                      </div>
-                    )}
-                  </div>
-                </fieldset>
-              </section>
+            {(derived.isStep1Locked || derived.isSynced) ? (
+              <StandardArticleFeaturedImagePanel
+                stagedArticle={stagedArticle}
+                isSynced={derived.isSynced}
+                isStep2Locked={derived.isStep2Locked}
+                featuredImagePreviewUrl={featuredImagePreviewUrl}
+                featuredImageTriggerLabel={featuredImageTriggerLabel}
+                onUpdateArticle={actions.setStageArticle}
+                onContinue={actions.continueFeaturedImage}
+                onOpenFeaturedImageModal={sidebarProps.onOpenFeaturedImageModal}
+              />
             ) : null}
 
-            {isStep1Locked || isSynced ? (
-              <section className="stl-panel sab-stage-panel sab-stage-panel-content">
-                <div className="stl-panel-header">
-                  <h2>{!isSynced ? <span className="stl-kicker">Step 3</span> : null} Content Blocks</h2>
-                  <div className="stl-inline-actions">
-                    <button type="button" className="stl-btn stl-btn-secondary" onClick={layout.onResetToOriginalBlocks}>
-                      Reset to Original
-                    </button>
-                    <button type="button" className="stl-btn stl-btn-secondary" onClick={handleOpenDeepExpand}>
-                      Deep Expand
-                    </button>
-                    {!isSynced ? (
-                      isStep3Locked ? (
-                        stagedArticle.step3_in_update_mode ? (
-                          <>
-                            <button type="button" className="stl-btn stl-btn-secondary" onClick={() => setStageArticle({ step3_in_update_mode: false })}>
-                              Cancel
-                            </button>
-                            <button type="button" className="stl-btn" onClick={handleContinueContent}>
-                              Save Step 3
-                            </button>
-                          </>
-                        ) : (
-                          <button type="button" className="stl-btn stl-btn-secondary" onClick={() => setStageArticle({ step3_in_update_mode: true })}>
-                            Update Step 3
-                          </button>
-                        )
-                      ) : (
-                        <button type="button" className="stl-btn" onClick={handleContinueContent}>
-                          Continue to Step 4
-                        </button>
-                      )
-                    ) : null}
-                  </div>
-                </div>
-
-                <div className="sab-stage-content-shell">
-                  <EditorialTimelineList {...timelineListProps} />
-                </div>
-              </section>
+            {(derived.isStep1Locked || derived.isSynced) ? (
+              <StandardArticleContentPanel
+                stagedArticle={stagedArticle}
+                timelineListProps={timelineListProps}
+                isSynced={derived.isSynced}
+                isStep3Locked={derived.isStep3Locked}
+                onUpdateArticle={actions.setStageArticle}
+                onContinue={actions.continueContent}
+                onResetToOriginalBlocks={layout.onResetToOriginalBlocks}
+                onOpenDeepExpand={expansion.open}
+              />
             ) : null}
 
-            {(isStep1Locked && isStep3Locked) || isSynced ? (
+            {((derived.isStep1Locked && derived.isStep3Locked) || derived.isSynced) ? (
               <SeoEditorPanel
-                seoSection={seoSection}
+                seoSection={derived.seoSection}
                 setSeoSection={updateSeoSection}
-                onGenerateSeoWithAi={handleGenerateSeoWithAi}
-                isGeneratingSeoTarget={isGeneratingSeoTarget}
-                onGenerateSeoImageFromFeatured={handleGenerateSeoImageFromFeatured}
-                isGeneratingSeoImage={isGeneratingSeoImage}
-                onUploadOgImageFile={handleUploadOgImageFile}
-                isUploadingOgImage={isUploadingOgImage}
-                onAutoFillOgUrl={handleAutoFillOgUrl}
+                onGenerateSeoWithAi={seo.generateSeo}
+                isGeneratingSeoTarget={seo.isGeneratingSeoTarget}
+                onGenerateSeoImageFromFeatured={seo.generateImageFromFeatured}
+                isGeneratingSeoImage={seo.isGeneratingSeoImage}
+                onUploadOgImageFile={seo.uploadOgImageFile}
+                isUploadingOgImage={seo.isUploadingOgImage}
+                onAutoFillOgUrl={permalink.autoFillOgUrl}
                 title="SEO & Sync"
               />
             ) : null}
           </main>
 
-          <aside className="stl-builder-sidebar">
-            {isSynced ? (
-              <section className="stl-summary-card">
-                <h3>Article Status</h3>
-                {syncIssues.length > 0 ? (
-                  <ul className="stl-summary-list">
-                    {syncIssues.map((issue) => <li key={issue}>{issue}</li>)}
-                  </ul>
-                ) : (
-                  <p className="stl-summary-note">All fields complete.</p>
-                )}
-              </section>
-            ) : (
-              <section className="stl-summary-card">
-                <h3>Build Progress</h3>
-                <div className="stl-progress-track" aria-hidden="true">
-                  <span className="stl-progress-bar" style={{ width: `${completionPercent}%` }} />
-                </div>
-                <p className="stl-summary-percent">{completionPercent}% ready</p>
-                <ul className="stl-summary-list">
-                  <li className={isStep1Locked ? 'done' : ''}>
-                    Setup: {isStep1Locked ? 'Locked' : stagedArticle.in_update_mode ? 'Editing' : 'Incomplete'}
-                  </li>
-                  <li className="done">
-                    Featured image: {isStep2Locked ? 'Saved' : stagedArticle.step2_in_update_mode ? 'Editing' : 'Skippable'}
-                  </li>
-                  <li className={isStep3Locked ? 'done' : ''}>
-                    Content blocks: {isStep3Locked ? 'Locked' : stagedArticle.step3_in_update_mode ? 'Editing' : isStep1Locked ? 'Ready' : 'Blocked'}
-                  </li>
-                  <li className={seoCoreComplete ? 'done' : ''}>
-                    SEO core: {seoCoreComplete ? 'Complete' : 'Missing SEO title or meta description'}
-                  </li>
-                </ul>
-              </section>
-            )}
-
-            <section className="stl-summary-card stl-summary-card--quick-actions">
-              <h3>Sync</h3>
-              <div className="stl-summary-actions">
-                {!isPublished ? (
-                  <button
-                    type="button"
-                    className="stl-btn stl-btn-success"
-                    onClick={() => handlePayloadSubmit('draft')}
-                    disabled={sidebarProps.isPublishing || syncIssues.length > 0}
-                  >
-                    {sidebarProps.isPublishing ? 'Saving...' : 'Save Draft to Payload'}
-                  </button>
-                ) : null}
-
-                <button
-                  type="button"
-                  className="stl-btn"
-                  onClick={() => handlePayloadSubmit('published')}
-                  disabled={sidebarProps.isPublishing || syncIssues.length > 0 || !canManagePublished}
-                >
-                  {sidebarProps.isPublishing
-                    ? 'Publishing...'
-                    : isPublished
-                      ? 'Update Published'
-                      : 'Publish'}
-                </button>
-              </div>
-
-              {stagedArticle.payloadArticleId ? (
-                <p className="stl-summary-note">
-                  {isPublished
-                    ? `Published Payload article: #${stagedArticle.payloadArticleId}`
-                    : `Linked Payload draft: #${stagedArticle.payloadArticleId}`}
-                </p>
-              ) : (
-                <p className="stl-summary-note">First sync will create a draft article in Payload.</p>
-              )}
-
-              {!canManagePublished ? (
-                <p className="stl-summary-note">
-                  Publishing requires an editor or admin role (you are signed in as {role ?? 'unknown'}).
-                </p>
-              ) : null}
-
-              {sidebarProps.publishResult ? (
-                <div className={`sab-stage-sync-result ${sidebarProps.publishResult.success ? 'success' : 'error'}`}>
-                  {sidebarProps.publishResult.message}
-                </div>
-              ) : null}
-
-              {syncIssues.length > 0 ? (
-                <div className="stl-summary-warning">
-                  <strong>Sync needs attention:</strong>
-                  <p>{syncIssues[0]}</p>
-                </div>
-              ) : null}
-            </section>
-
-            <section className="stl-summary-card">
-              <h3>Draft Details</h3>
-              <div className="sab-stage-meta-list">
-                <p><strong>Run ID:</strong> {stagedArticle.runId || 'n/a'}</p>
-                <p><strong>Location:</strong> {selectedLocationLabel || 'Not set'}</p>
-                {stagedArticle.createdBy ? (
-                  <p><strong>Created by:</strong> {stagedArticle.createdBy.name || stagedArticle.createdBy.email}</p>
-                ) : null}
-                {stagedArticle.lastEditedBy ? (
-                  <p><strong>Last edited by:</strong> {stagedArticle.lastEditedBy.name || stagedArticle.lastEditedBy.email}</p>
-                ) : null}
-                <p><strong>Created:</strong> {stagedArticle.createdAt ? new Date(stagedArticle.createdAt).toLocaleDateString() : 'n/a'}</p>
-                <p><strong>Updated:</strong> {stagedArticle.updatedAt ? new Date(stagedArticle.updatedAt).toLocaleDateString() : 'n/a'}</p>
-              </div>
-            </section>
-          </aside>
+          <StandardArticleBuilderSidebar
+            stagedArticle={stagedArticle}
+            sidebarProps={sidebarProps}
+            selectedLocationLabel={derived.selectedLocationLabel}
+            syncIssues={derived.syncIssues}
+            completionPercent={derived.completionPercent}
+            isSynced={derived.isSynced}
+            isPublished={derived.isPublished}
+            isStep1Locked={derived.isStep1Locked}
+            isStep2Locked={derived.isStep2Locked}
+            isStep3Locked={derived.isStep3Locked}
+            seoCoreComplete={derived.seoCoreComplete}
+            canManagePublished={canManagePublished}
+            role={role}
+            onSubmit={actions.submitToPayload}
+          />
         </div>
       </div>
 
       <FeaturedImageModal {...featuredModalProps} />
       <BlockImageModal {...blockModalProps} />
-
       <ArticleExpansionModal
-        isOpen={isExpansionModalOpen}
+        isOpen={expansion.isOpen}
         phase={expansion.phase}
         expandedArticle={expansion.expandedArticle}
         expansionPlan={expansion.expansionPlan}
         error={expansion.error}
-        originalArticle={stagedArticle?.content ?? ''}
-        articleType={stagedArticle?.originalType ?? ''}
-        title={stagedArticle?.title || stagedArticle?.originalTitle || ''}
+        originalArticle={stagedArticle.content}
+        articleType={stagedArticle.originalType}
+        title={stagedArticle.title || stagedArticle.originalTitle}
         listicleDetection={expansion.listicleDetection}
-        onAccept={handleAcceptExpansion}
-        onClose={handleCloseExpansion}
+        onAccept={expansion.accept}
+        onClose={expansion.close}
         onProceed={expansion.proceedWithExpansion}
       />
     </>
